Parsing agreement encapsulated into GetOrCreateJob
